move logger + refactoring
