+++ bluetooth/src/commonMain/kotlin/Descriptor.kt
@@ -19,6 +19,7 @@ package com.splendo.kaluga.bluetooth
 
 import com.splendo.kaluga.bluetooth.device.DeviceAction
 import com.splendo.kaluga.bluetooth.device.DeviceConnectionManager
+import com.splendo.kaluga.logging.ContextualLogger
 
 /**
  * An [Attribute] of a Bluetooth Descriptor
