+++ bluetooth/src/commonMain/kotlin/Service.kt
@@ -18,6 +18,7 @@
 package com.splendo.kaluga.bluetooth
 
 import com.splendo.kaluga.bluetooth.device.DeviceConnectionManager
+import com.splendo.kaluga.logging.ContextualLogger
 
 /**
  * A Bluetooth Service
