+++ test-utils-bluetooth/src/commonMain/kotlin/MockDescriptor.kt
@@ -17,7 +17,7 @@
 
 package com.splendo.kaluga.test.bluetooth
 
-import com.splendo.kaluga.bluetooth.ContextualLogger
+import com.splendo.kaluga.logging.ContextualLogger
 import com.splendo.kaluga.bluetooth.Descriptor
 import com.splendo.kaluga.bluetooth.DescriptorWrapper
 import com.splendo.kaluga.bluetooth.device.DeviceConnectionManager
