+++ test-utils-bluetooth/src/commonMain/kotlin/device/MockDevice.kt
@@ -17,7 +17,7 @@
 
 package com.splendo.kaluga.test.bluetooth.device
 
-import com.splendo.kaluga.bluetooth.ContextualLogger
+import com.splendo.kaluga.logging.ContextualLogger
 import com.splendo.kaluga.bluetooth.RSSI
 import com.splendo.kaluga.bluetooth.Service
 import com.splendo.kaluga.bluetooth.device.BaseAdvertisementData
